refactor: tighten offset-regime area after fix for #920
Repoint the layout-layer apply_route_offsets imports from render.svg to its
canonical home routing/common.py (the renderer still re-exports it for render
code and tests), so layout passes no longer reach up into the render layer for
a layout utility. Document the shared lateral-offset tolerance.
